Get the client running: six mixins that only fail when applied
The remapper's inventory only catches targets whose name moved. These six
remapped clean and still failed on apply, because what moved was the
injection point or the callback type -- and for a screen class Mixin only
finds that out when the screen first opens.

- Minecraft.pickBlock was rewritten around handlePickItemFrom*; the
  Abilities.instabuild read the injection hung on no longer exists.
- Minecraft.startUseItem lost InteractionResult.shouldSwing(); the check
  is now Success.swingSource() == CLIENT. Returning SERVER instead skips
  the whole branch, which is what setting shouldSwing false used to do.
- SoundEngine.play returns PlayResult, so the callback is a CIR.
- ItemInHandRenderer hands out a SubmitNodeCollector now. That runs
  through RenderHandEvent, IFPGeoItemRenderer and both first-person
  renderers, none of which used the buffer source anyway -- the stencil
  path draws immediately on purpose.
- GameRenderer.getFov returns float. The event stays on double like
  upstream; the conversion happens at the mixin boundary.
- CreateWorldScreen.openFresh only forwards; the PackRepository is built
  in openCreateWorldScreen. This one was load-bearing: SelectWorldScreen
  links against CreateWorldScreen, so the failure took the world list
  down with it -- an empty panel, and --quickPlaySingleplayer hanging on
  "Loading Minecraft" with nothing in the log.

Also: drop Iris from the runtime classpath. Sodium 0.8.13 declares
`breaks iris <=1.10.7` and 1.10.7 is Iris's newest 1.21.11 release, so
the two current versions are mutually exclusive upstream and the dev
client stops at mod resolution. Iris stays modCompileOnly; the shader
compat code just does not execute in dev until Iris ships a fix.

And a -PquickPlay=<save> hook on the client run config, to skip the menu
when checking rendering over and over.

Verified in game: title screen, world list, loading a world, the
first-person gun with the player's arm, the third-person gun, the HUD,
creative inventory with REI, tooltips, chat, iron-sight aiming.

The scope stencil mask is half good news. The stencil buffer itself
works -- the inverted circle cuts a real hole, which it could not do if
the attachment were missing -- but the mask lands wrong: aiming an
optical scope paints the screen black and shows the world in a ring
around the tube. Iron sights are unaffected. Roadmap records where to
look next.

+++ src/main/java/com/tacz/guns/mixin/client/GameRendererMixin.java
@@ -75,7 +75,7 @@ public abstract class GameRendererMixin {
      * 至于为什么不直接对 renderItemInHand 这个方法 mixin ，是因为安装了 Optifine 之后，这个方法的内容被大幅度修改了。
      */
     @Inject(method = "getFov", at = @At("HEAD"))
-    public void switchRenderType(Camera pActiveRenderInfo, float pPartialTicks, boolean pUseFOVSetting, CallbackInfoReturnable<Double> cir) {
+    public void switchRenderType(Camera pActiveRenderInfo, float pPartialTicks, boolean pUseFOVSetting, CallbackInfoReturnable<Float> cir) {
         this.tacz$useFovSetting = pUseFOVSetting;
     }
 
@@ -84,11 +84,12 @@ public abstract class GameRendererMixin {
      * SimpleBedrockModel, which patched the same method from its own mixin.
      */
     @ModifyReturnValue(method = "getFov", at = @At(value = "RETURN", ordinal = 1))
-    private double tacz$computeFov(double original, @Local(argsOnly = true) Camera camera,
-                                   @Local(argsOnly = true) float partialTicks,
-                                   @Local(argsOnly = true) boolean useConfigured) {
+    // getFov 现在返回 float；事件本身仍按上游那样用 double，只在这个边界上转一次
+    private float tacz$computeFov(float original, @Local(argsOnly = true) Camera camera,
+                                  @Local(argsOnly = true) float partialTicks,
+                                  @Local(argsOnly = true) boolean useConfigured) {
         ViewportEvent.ComputeFov event = new ViewportEvent.ComputeFov((GameRenderer) (Object) this, camera, partialTicks, original, useConfigured);
         ViewportEvent.FOV.invoker().post(event);
-        return event.getFOV();
+        return (float) event.getFOV();
     }
 }
